backend/pyproject.toml
[project]
name = "fine-audit-backend"
version = "0.1.0"
description = "Auditoría automática de modelos ML (backend)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Equipo Fine", email = "dev@example.com" }]
dependencies = [
  "fastapi>=0.115.0,<0.116.0",
  "uvicorn[standard]>=0.30.0,<0.31.0",
  "pydantic>=2.9.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "aiosqlite>=0.19.0,<0.20.0",
  "asyncpg>=0.29.0,<0.30.0",
  "psycopg2-binary>=2.9.9,<3.0.0",
  "sqlmodel>=0.0.22,<0.0.23",
  "sqlalchemy>=2.0.30,<3.0.0",
  "alembic>=1.13.2,<2.0.0",
  "redis>=5.0.7,<6.0.0",
  "celery>=5.4.0,<6.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "orjson>=3.10.0,<4.0.0",
  "numpy>=1.26.4,<2.0.0",
  "pandas>=2.2.2,<3.0.0",
  "scikit-learn>=1.5.1,<1.6.0",
  "xgboost>=2.1.1,<2.2.0",
  "torch>=2.3.1,<2.4.0",
  "shap>=0.45.1,<0.46.0",
  "langchain==0.3.12",
  "langgraph>=0.2.0,<0.3.0",
  "langchain-openai==0.2.0",
  "langfuse>=2.0.0,<3.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "boto3>=1.35.0,<1.36.0",
  "prometheus-client>=0.20.0,<0.21.0",
  "openai>=1.44.0,<2.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "weasyprint>=62.3,<63.0",
  "httpx>=0.27.0,<0.28.0",
  "structlog>=24.4.0,<25.0.0"
]

[project.optional-dependencies]
fairness = ["aif360>=0.6.1,<0.7.0"]
dev = [
  "ruff>=0.6.2,<0.7.0",
  "black>=24.8.0,<25.0.0",
  "mypy>=1.11.2,<1.12.0",
  "pytest>=8.3.2,<9.0.0",
  "pytest-asyncio>=0.23.8,<0.24.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "scikit-learn>=1.5.1,<1.6.0",
  "pandas>=2.2.2,<3.0.0",
  "numpy>=1.26.4,<2.0.0",
  "joblib>=1.4.2,<2.0.0",
  "pre-commit>=3.8.0,<4.0.0",
  "types-redis>=4.6.0.20240724"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "B", "UP", "C4", "SIM", "RUF"]
lint.ignore = ["E501"]
format.quote-style = "double"
format.indent-style = "space"
format.skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = false
disallow_untyped_calls = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
namespace_packages = true

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.ruff.lint.isort]
known-first-party = ["app"]

